Describe the output format for procdoc.awk
Given that the intended output is a set of JSON objects, it makes sense
to define exactly what this output is via a JSON schema.

This will probably be extended to include enum/macro/struct/union
documentation at some point, hence the "function" content using the term
"members" for function parameters.